Saas.Pushover

Description

Small library that provides functions to send push messages to Android and Apple devices which have the Pushover app installed.

Data types

data PushMessage

The PushMessage data structure. To construct one of these, you should alter the message under *defaultMessage* using record syntax.

Constructors

PM

Fields

token :: ByteString

user :: ByteString

message :: ByteString

device :: ByteString

title :: ByteString

url :: ByteString

urlTitle :: ByteString

priority :: Int

timestamp :: ByteString

sound :: ByteString

callback :: ByteString

expire :: Int

retry :: Int

data PushResponse

When you send a PushMessage, the server replies with at least a status code and a request number. See the pushover API documentation for what each field means.

Constructors

PR

Fields

status :: Int

request :: ByteString

receipt :: Maybe ByteString

errors :: Maybe [ByteString]

data ReceiptResponse

The reponse when you inquire about a receipt for a priority 2 message. See the pushover API documentation for what each field means.

Constructors

RR

Fields

receiptstatus :: Int

acknowledged :: Int

acknowledgedAt :: Int

lastDeliveredAt :: Int

expired :: Int

expiresAt :: Int

calledBack :: Int

calledBackAt :: Int

type Apptoken = ByteString

type Receipt = ByteString

Default constructor

defaultMessage :: PushMessage

A default PushMessage (all empty fields except token, user and message will be removed later in the POST request, but the fields have to be there to overwrite them later (if you want)).

IO functions

sendPushMessage :: PushMessage -> IO PushResponse

Sends a push message to the Pushover servers.

checkReceipt :: Apptoken -> Receipt -> IO ReceiptResponse

Inquire about a receipt.